+++ graphql_client/src/queries.ts
@@ -39,6 +39,7 @@ export const GET_META_QUERY = `
 				validation
 				cardinality
 				source
+				config
 			}
 			workflow {
 				states
@@ -106,6 +107,7 @@ export const GET_ALL_META_QUERY = `
 				validation
 				cardinality
 				source
+				config
 			}
 			workflow {
 				states